Added one question and some minor improvements for better understanding. Additionally fixed the array check in klicktipp-tag-emails.js and removed unused code. Also renamed file to fix typo
