+++ nodes/klicktipp-login.html
@@ -23,7 +23,7 @@
 	   <code>msg.payload</code> will contain <code>{ success: false }</code>.
 	</p>
 	<p>
-	   Upon successful login, an automatic logout is scheduled and will occur after 10 minutes unless canceled manually.
+	   Upon successful login, an automatic logout is scheduled.
 	</p>
 
 	<h3>Automatic Logout</h3>
