+++ nodes/klicktipp-signin.html
@@ -39,8 +39,8 @@
 		   Must contain:
 		   <ul>
 			   <li><code>apiKey</code> - (Required) The API key for authentication.</li>
-			   <li><code>email</code> - (Optional) The email address of the subscriber to sign in.</li>
-			   <li><code>smsNumber</code> - (Optional) The SMS number of the subscriber to sign in (if email is not provided).</li>
+			   <li><code>email</code> - (Optional) The email address of the subscriber to sign in (mandatory if no SMS number is provided).</li>
+			   <li><code>smsNumber</code> - (Optional) The SMS number of the subscriber to sign in (mandatory if email is not provided).</li>
 			   <li><code>fields</code> - (Optional) Additional fields for custom data (e.g., name, other details).</li>
 		   </ul>
 	   </dd>
