+++ nodes/klicktipp-subscription-process-get.html
@@ -32,7 +32,7 @@
 	<dd>
 	   An object that must contain the following:
 	   <ul>
-		   <li><code>listId</code> - TThe ID of the subscription process (list) to retrieve.</li>
+		   <li><code>listId</code> - The ID of the subscription process (list) to retrieve.</li>
 	   </ul>
 	</dd>
 	</dl>
