+++ nodes/utils/generateUniqueSessionKey.js
@@ -3,6 +3,8 @@
  *
  * @returns {string} A unique session key for storing session data.
  */
+
+// question: why not use a UUID library?
 function generateUniqueSessionKey() {
 	const timestamp = Date.now();
 	const randomNum = Math.floor(Math.random() * 1000000);
